Create separate folders for different components
