+++ src/App.js
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { Routes, Route } from 'react-router-dom';
-import Missions from './components/Missions';
-import MyProfile from './components/MyProfile';
+import Missions from './components/Missions/Missions';
+import MyProfile from './components/MyProfile/MyProfile';
 import NavLink from './components/NavLink';
 import { fetchMission } from './components/redux/missions/missions';
 import Rocket from './components/Rocket';
@@ -12,6 +12,7 @@ function App() {
   useEffect(() => {
     dispatch(fetchMission());
   }, []);
+
   return (
     <>
       <NavLink />
