+++ src/components/Missions/missions.css
@@ -1,0 +1,15 @@
+.tableMissions,
+th,
+td {
+  border: 1px solid black;
+  border-collapse: collapse;
+}
+
+.reservation {
+  color: black;
+  padding: 10px;
+  text-decoration: none;
+  cursor: pointer;
+  border-radius: 8px;
+  border: 1px solid black;
+}
